Initial commit: create basic org-mode file for tangling
Since I am re-creating an Emacs configuration from scratch, I want to
use the idea that's been playing around in my mind for ages:
Maintaining the full configuration in an org-mode file and generating
the configuration from this file.

I am using Prot's org-mode file as a reference file, so the initial
work will be lots of copying from his work. Once the skeleton is in
place, I will add all of my custom configuration to this file as well.