Load emacs-pet from github
Finally, this commit puts elpaca to use and does the thing I've been
wanting to do, which is being able to develop packages that I am using
in a seamless way and being able to contribute PRs upstream.

If this experiment goes well, I will merge this branch back into main
and continue using elpaca for my package management needs.